refactor: clean up breadcrumbs component
