+++ packages/breadcrumbs/docs/Breadcrumbs.mdx
@@ -40,6 +40,6 @@ attribute set. Usually this is the last crumb in the trail.
 
 ## Props
 
-```props packages/breadcrumbs/src/Breadcrumbs.tsx
+```props packages/breadcrumbs/src/component.tsx
 
 ```
